+++ tests/task/test_calcZernikesUnpairedTask.py
@@ -233,6 +233,9 @@ class TestCalcZernikeUnpaired(lsst.utils.tests.TestCase):
                     "MAX_POWER_GRAD_SELECT",
                     "FINAL_SELECT",
                     "DEFOCAL_TYPE",
+                    "RADIUS",
+                    "X_PIX_LEFT_EDGE",
+                    "X_PIX_RIGHT_EDGE"
                 ]
                 np.testing.assert_array_equal(
                     np.sort(colnames), np.sort(desired_colnames)